fix narrowing: change UbUtilEntry.upperBound to double; cleanup redundant (void)rankSize

+++ src/ops/op_common/selector/cost_table.h
@@ -32,7 +32,7 @@ typedef struct {
 } CostTable;
 
 struct UbUtilEntry {
-    u64 upperBound;
+    double upperBound;
     float utilization;
 };
